moves header component into components folder
